refactor: reorganize main.go into focused packages and files
Split the large main.go file (614 lines) into focused packages:

**New Package Structure:**
- `internal/app/` - Application configuration and types
  * `config.go` - Environment setup and client initialization
  * `types.go` - Core data structures (SheetRow, SheetRowUpdate, SheetItem)

- `internal/resolution/` - User and item resolution utilities
  * `users.go` - User lookup, matching, and fallback logic
  * `items.go` - Item lookup, matching, market value, and fallback logic

- `internal/sheets/` - Google Sheets operations
  * `operations.go` - Sheet reading, parsing, and appending
  * `updates.go` - Sheet update operations for provided items
  * `config.go` - Environment variable utilities

- `internal/processing/` - Business logic processing
  * `supplied.go` - Supplied items processing and sheet row generation
  * `provided.go` - Provider matching and update logic

**Improvements:**
- **Separation of Concerns**: Each package has a single, clear responsibility
- **Reduced Complexity**: Main.go reduced from 614 lines to 85 lines
- **Better Testability**: Smaller, focused functions can be unit tested independently
- **Clear Dependencies**: Import structure makes relationships explicit
- **Eliminated Import Cycles**: Careful dependency management prevents circular imports
- **Maintainability**: Related functionality grouped together
- **Scalability**: Easy to add new operations or processing types

**Main.go Now Contains:**
- Application entry point and orchestration
- Main processing loop coordination
- Provider list management

All business logic has been extracted to appropriate packages while maintaining the exact same functionality.
